first pass reorganizing directory/file structure
