+++ .gitignore
@@ -140,4 +140,4 @@ dmypy.json
 # Cython debug symbols
 cython_debug/
 
-!/asb.config
+!/config/asb.config
