+++ airspotbot/__main__.py
@@ -1,0 +1,3 @@
+from .airspotbot import run_bot
+run_bot()
+
