+++ airspotbot/location.py
